tests/requirements.txt
pact-python
pytest
unitest
requests
os